fixed more cqlsh_copy_tests

+++ pylib/cqlshlib/wcwidth.py
@@ -85,7 +85,7 @@ def bisearch(ucs, table):
     if ucs < table[0][0] or ucs > table[max][1]:
         return 0
     while max >= min:
-        mid = (min + max) / 2
+        mid = int((min + max) / 2)
         if ucs > table[mid][1]:
             min = mid + 1
         elif ucs < table[mid][0]:
